modified fils sort & rename
