+++ light-life/package.json
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:5e22c144f3c255467bc395a51a3f51c3bc52c66c280e642d3df3aae94278ec15
-size 1397
+oid sha256:97a3a92f4be65c05e672b07f6f011ec97ad309dd26b495b6a93e2ccb9fc3a0d0
+size 1481
