+++ light-life/public/index.html
@@ -15,7 +15,7 @@
     />
     <script src="https://use.fontawesome.com/28ff5fa4b5.js"></script>
 
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
+    <link rel="apple-touch-icon" href="../src/images/icon.svg" />
 
     <!--
       manifest.json provides metadata used when your web app is installed on a
