+++ light-life/src/Pages/Components/DietaryRecord/DietaryRecord.js
@@ -1,0 +1,155 @@
+import React, { useEffect, useState } from "react";
+import firebase from "firebase/app";
+import {
+  BrowserRouter as Router,
+  Switch,
+  Route,
+  Link,
+  useLocation,
+  useParams,
+} from "react-router-dom";
+import "firebase/firestore";
+import DietitianRecord from "./DietitianRecord.js";
+import CustomerRecord from "./CustomerRecord.js";
+// import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+// import { faTimes } from "@fortawesome/free-solid-svg-icons";
+
+// function ShowImages({ mealDetails, input, removeImageHandler }) {
+//   if (!input.images) {
+//     return (
+//       <>
+//         {mealDetails.images.map((i, index) => (
+//           <div key={index}>
+//             <div id={index} onClick={removeImageHandler}>
+//               X
+//             </div>
+//             <a href={i} target="_blank" rel="noreferrer noopener">
+//               <img
+//                 src={i}
+//                 alt="customer"
+//                 style={{ width: "200px", height: "200px" }}
+//               />
+//             </a>
+//           </div>
+//         ))}
+//         {input.imageUrl.map((i, index) => (
+//           <div key={index}>
+//             <a href={i} target="_blank" rel="noreferrer noopener">
+//               <img
+//                 src={i}
+//                 alt="customer"
+//                 style={{ width: "200px", height: "200px" }}
+//               />
+//             </a>
+//           </div>
+//         ))}
+//       </>
+//     );
+//   } else {
+//     return (
+//       <>
+//         {input.images.map((i, index) => (
+//           <div key={index}>
+//             <div id={index} onClick={removeImageHandler}>
+//               X
+//             </div>
+//             <a href={i} target="_blank" rel="noreferrer noopener">
+//               <img
+//                 src={i}
+//                 alt="customer"
+//                 style={{ width: "200px", height: "200px" }}
+//               />
+//             </a>
+//           </div>
+//         ))}
+//         {input.imageUrl.map((i, index) => (
+//           <div key={index}>
+//             <a href={i} target="_blank" rel="noreferrer noopener">
+//               <img
+//                 src={i}
+//                 alt="customer"
+//                 style={{ width: "200px", height: "200px" }}
+//               />
+//             </a>
+//           </div>
+//         ))}
+//       </>
+//     );
+//   }
+// }
+
+function RenderDietaryRecord() {
+  const [recordDate, setRecordDate] = useState();
+  const [getRecord, setGetRecord] = useState(false); //false
+  const [count, setCount] = useState(1);
+  const params = useParams();
+
+  const getDietaryRecordDate = (e) => {
+    if (e.target.value !== "") {
+      setRecordDate(e.target.value);
+      setCount(1);
+      setGetRecord(true);
+    }
+  };
+  if (params.dID) {
+    return (
+      <>
+        <input
+          type="date"
+          min="2021-05-14"
+          max="2021-05-26"
+          onChange={getDietaryRecordDate}
+          required="required"
+        ></input>
+        <Router>
+          <Link
+            to={`/dietitian/${params.dID}/customer/${params.cID}/dietary/`}
+          ></Link>
+          {getRecord ? (
+            <Switch>
+              <Route exact path={`/dietitian/:dID/customer/:cID/dietary/`}>
+                <DietitianRecord
+                  date={recordDate}
+                  count={count}
+                  setCount={setCount}
+                  style={{ boxSizing: "border-box" }}
+                />
+              </Route>
+            </Switch>
+          ) : (
+            ""
+          )}
+        </Router>
+      </>
+    );
+  } else {
+    return (
+      <>
+        <input
+          type="date"
+          min="2021-05-14"
+          max="2021-05-31"
+          onChange={getDietaryRecordDate}
+        ></input>
+        <Router>
+          <Link to={`/customer/${params.cID}/dietary/`}></Link>
+          {getRecord ? (
+            <Switch>
+              <Route exact path="/customer/:cID/dietary/">
+                <CustomerRecord
+                  date={recordDate}
+                  count={count}
+                  setCount={setCount}
+                />
+              </Route>
+            </Switch>
+          ) : (
+            ""
+          )}
+        </Router>
+      </>
+    );
+  }
+}
+
+export default RenderDietaryRecord;
