--- light-life/src/Pages/Dietitian/ingredientSearch.js
@@ -1,2 +1,0 @@
-// import React, { useEffect, useState } from "react";
-// import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
